Rename foder on component
