+++ react-basic-practice/src/components/Icons/DeleteIcon.tsx
