+++ react-basic-practice/src/components/Icons/EditIcon.tsx
