+++ react-basic-practice/src/components/Icons/ViewIcon.tsx
