+++ react-basic-practice/src/components/Modal/index.tsx
